Book class & some idea
Hey guys. I have now finished implementing the book class. Next step has
to be "how will admin set the book loan statuses". There is a boolean
isOnLoan in the book class. So where will the admin change this? How
will the program make a book on loan to a student. Will it be in a class
named AdminTransaction or, or will it be just a class that interacts
with the database and reflects the loan status & helps the admin change
it? Please share ideas